TAllocator, TMemoryDeleter, PCH file for OpenBlob Project TAllocator helps STL to use TMemory to allocate and free memory Renamed Toshi's PCH to not conflict with OpenBlob's new PCH Renamed TVectorNum to TMathFloating Added TKernelInterface::GetAverageFPS

+++ Toshi/Source/Toshi/Core/TObject.h
@@ -1,8 +1,8 @@
 #pragma once
-#include "Core.h"
 #include "TClass.h"
-#include "TMemory.h"
 
+#include "Toshi/Core/Core.h"
+#include "Toshi/Core/TMemory.h"
 #include "Toshi/Utils/TLog.h"
 
 #define TOSHI_CLASS_DEFINE(CLASSNAME) \
@@ -42,7 +42,7 @@ namespace Toshi
 		virtual ~TObjectInterface();
 
 		// Operators
-		static inline void* operator new(size_t size) { return tmalloc((uint32_t)size); }
+		static inline void* operator new(size_t size) { return tmalloc(size); }
 		static inline void* operator new(size_t size, void* at) { return at; }
 
 		static inline void operator delete(void* block) { tfree(block); }
